+++ src/test/kotlin/com/pestphp/pest/types/ExpectCallCompletionTest.kt
@@ -1,0 +1,61 @@
+package com.pestphp.pest.types
+
+import com.intellij.openapi.components.service
+import com.intellij.testFramework.replaceService
+import com.jetbrains.php.composer.lib.ComposerLibraryManager
+import com.jetbrains.php.lang.psi.PhpFile
+import com.pestphp.pest.services.ExpectationFileService
+import org.mockito.kotlin.doReturn
+import org.mockito.kotlin.mock
+import java.util.concurrent.TimeUnit
+
+class ExpectCallCompletionTest : BaseTypeTest() {
+    override fun setUp() {
+        super.setUp()
+
+        val dir = myFixture.copyDirectoryToProject("expect", "tests")
+        myFixture.copyFileToProject("stubs.php")
+
+        val composerMock = mock<ComposerLibraryManager> {
+            on { findVendorDirForUpsource() } doReturn dir
+        }
+        project.replaceService(
+            ComposerLibraryManager::class.java,
+            composerMock,
+            testRootDisposable
+        )
+    }
+
+    fun testFieldCompletions() {
+        val file = myFixture.configureByFile("tests/expectCallCompletion.php")
+
+        val service = project.service<ExpectationFileService>()
+        service.updateExtends(file as PhpFile)
+        service.generateFile()
+        waitForAppLeakingThreads(10, TimeUnit.SECONDS)
+
+        assertCompletion("someExtend")
+    }
+
+    fun testFieldCompletionsChainedAndProperty() {
+        val file = myFixture.configureByFile("tests/expectCallCompletionChainedAndProperty.php")
+
+        val service = project.service<ExpectationFileService>()
+        service.updateExtends(file as PhpFile)
+        service.generateFile()
+        waitForAppLeakingThreads(10, TimeUnit.SECONDS)
+
+        assertCompletion("someExtend")
+    }
+
+    fun testFieldCompletionsChainedAndMethod() {
+        val file = myFixture.configureByFile("tests/expectCallCompletionChainedAndMethod.php")
+
+        val service = project.service<ExpectationFileService>()
+        service.updateExtends(file as PhpFile)
+        service.generateFile()
+        waitForAppLeakingThreads(10, TimeUnit.SECONDS)
+
+        assertCompletion("someExtend")
+    }
+}