+++ src/test/resources/com/pestphp/pest/types/expect/expectCallCompletion.php
@@ -1,0 +1,10 @@
+<?php
+
+expect()->extend('someExtend', function ($min, $max) {
+    return $this->toBeGreaterThanOrEqual($min)
+        ->toBeLessThanOrEqual($max);
+});
+
+test('numeric ranges', function () {
+    expect(100)-><caret>
+});